feat:
- add favicon
- update rootlayout
